requirements.txt
alembic==1.13.1
bcrypt==4.1.2
boto3==1.34.68
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
gunicorn==21.2.0
passlib==1.7.4
psycopg2-binary==2.9.9
py3dns==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
SQLAlchemy==2.0.28
SQLAlchemy-Utils==0.41.1
starlette-admin==0.13.2
uvicorn==0.28.0
validate-email==1.3
pydantic-settings==2.2.1
filetype==1.2.0
fastapi-cache2==0.2.1
aioredis==2.0.1
asyncpg==0.29.0
sqladmin-0.16.1